move setValue function to CobolHeapPointer
