+++ src/main/java/com/github/lopoha/coboltruffle/nodes/expression/heap/CobolHeapPointer.java
@@ -6,6 +6,7 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 // todo: make class abstract and have multiple implementations: Number, String, ...
 // todo: should this class be under nodes/expressions insted of heap?
@@ -99,8 +100,27 @@ public abstract class CobolHeapPointer extends CobolExpressionNode {
    * The value is automatically aligned/cut/...
    * @param value the new value
    */
-  public abstract void setValue(List<Character> value, CobolProgramStateNode programStateNode);
-
+  public void setValue(List<Character> value, CobolProgramStateNode programState) {
+    assert value != null : "Value required but was null";
+    // todo: check if the alignment etc. is correct
+    // todo: is the default for number here space or zero?
+    // todo: is parallel always faster?
+    if (value.size() < this.length) {
+      final String val
+          = new String(new char[this.length - value.size()]).replace('\0', ' ') + value;
+      IntStream.range(0, value.size())
+          .parallel()
+          .forEach(i -> programState.getLocalFileHeap().set(position + i, val.charAt(i)));
+    } else if (value.size() == this.length) {
+      IntStream.range(0, value.size())
+          .parallel()
+          .forEach(i -> programState.getLocalFileHeap().set(position + i, value.get(i)));
+    } else {
+      IntStream.range(0, this.length)
+          .parallel()
+          .forEach(i -> programState.getLocalFileHeap().set(position + i, value.get(i)));
+    }
+  }
   /**
    * Compares the HeapPointer value to another HeapPointer valuie.
    * @param o the other HeapPointer.
