+++ src/main/java/com/github/lopoha/coboltruffle/nodes/expression/heap/CobolHeapPointerRaw.java
@@ -35,29 +35,6 @@ public final class CobolHeapPointerRaw extends CobolHeapPointer {
   }
 
   @Override
-  public void setValue(List<Character> value, CobolProgramStateNode programState) {
-    assert value != null : "Value required but was null";
-    // todo: check if the alignment etc. is correct
-    // todo: is the default for number here space or zero?
-    // todo: is parallel always faster?
-    if (value.size() < this.length) {
-      final String val
-          = new String(new char[this.length - value.size()]).replace('\0', ' ') + value;
-      IntStream.range(0, value.size())
-          .parallel()
-          .forEach(i -> programState.getLocalFileHeap().set(position + i, val.charAt(i)));
-    } else if (value.size() == this.length) {
-      IntStream.range(0, value.size())
-          .parallel()
-          .forEach(i -> programState.getLocalFileHeap().set(position + i, value.get(i)));
-    } else {
-      IntStream.range(0, this.length)
-          .parallel()
-          .forEach(i -> programState.getLocalFileHeap().set(position + i, value.get(i)));
-    }
-  }
-
-  @Override
   public int compareTo(CobolHeapPointer o) {
     // todo
     // return this.getValue(null).compareTo(o.getValue(null));
